refactor(automation): effects get own files
