+++ tests/cogs5e/models/automation/effects_test.py
@@ -105,7 +105,7 @@ class TestUseCustomCounter:
         }
         result = automation.UseCounter.from_data(data)
         assert result
-        assert isinstance(result.counter, automation.utils.SpellSlotReference)
+        assert isinstance(result.counter, automation.effects.usecounter.SpellSlotReference)
         assert result.counter.slot == 3
 
         data = {
@@ -115,26 +115,27 @@ class TestUseCustomCounter:
         }
         result = automation.UseCounter.from_data(data)
         assert result
-        assert isinstance(result.counter, automation.utils.AbilityReference)
+        assert isinstance(result.counter, automation.effects.usecounter.AbilityReference)
         assert result.counter.entity is compendium.lookup_entity(12168134, 75)
 
     async def test_serialize(self):
         result = automation.UseCounter('Bardic Inspiration', '1').to_dict()
         assert json.dumps(result)  # result should be JSON-encodable
 
-        result = automation.UseCounter(automation.utils.SpellSlotReference(1), '1').to_dict()
+        result = automation.UseCounter(automation.effects.usecounter.SpellSlotReference(1), '1').to_dict()
         assert json.dumps(result)
 
-        result = automation.UseCounter(automation.utils.AbilityReference(12168134, 75), '1').to_dict()
+        result = automation.UseCounter(automation.effects.usecounter.AbilityReference(12168134, 75), '1').to_dict()
         assert json.dumps(result)
 
     @pytest.mark.parametrize("counter", [
         # counters by name
         "counter", "other counter name",
         # spell slot
-        automation.utils.SpellSlotReference(1), automation.utils.SpellSlotReference(9),
+        automation.effects.usecounter.SpellSlotReference(1), automation.effects.usecounter.SpellSlotReference(9),
         # ability reference
-        automation.utils.AbilityReference(12168134, 75), automation.utils.AbilityReference(-1, -1)
+        automation.effects.usecounter.AbilityReference(12168134, 75),
+        automation.effects.usecounter.AbilityReference(-1, -1)
     ])
     @pytest.mark.parametrize("amount", [
         # valid inputs
